+++ src_emscripten/run_demo.cpp
@@ -9,17 +9,17 @@
 extern "C"
 {
 
-halide_buffer_t *run_demo(int32_t *data, int width, int height)
+halide_buffer_t *run_demo(uint8_t *data, int width, int height)
 {
     std::cout << "In C++" << std::endl;
 
     // Allocates halide_buf_wrapper on heap, to prevent the buffer from being deallocated at the
     // end of this function before the async code in halide_myfunc completes:
-    Halide::Runtime::Buffer<int32_t> *halide_buf_wrapper = new Halide::Runtime::Buffer<int32_t>{data, width, height};
+    Halide::Runtime::Buffer<uint8_t> *halide_buf_wrapper = new Halide::Runtime::Buffer<uint8_t>{data, width, height};
     halide_buffer_t *halide_buf = halide_buf_wrapper->raw_buffer();
     std::cout << "halide_buf address = " << halide_buf << std::endl;
 
-    int32_t *halide_data = get_halide_buffer_data(halide_buf);
+    uint8_t *halide_data = get_halide_buffer_data(halide_buf);
     std::cout << "halide_buf data address = " << halide_data << std::endl;
     std::cout << "halide_buf data in memory:" << std::endl;
     std::cout << halide_data[0];
